chore(#723): extract JSON logging to cli/_logging + document source-only connectors
Two small cleanups from the #723 grab-bag (the mcp/server.py tool split and the
schema.py table-drive are deferred — the latter collides with #707).

- **JSON logging** — `_JsonFormatter` / `_configure_json_logging` / the
  `_STANDARD_LOG_FIELDS` set move out of `cli/commands/run.py` into a dedicated
  `cli/_logging.py`. `run.py` imports the configure helper it uses; `cli/main.py`
  imports both from `_logging` for its back-compat re-export block, so the public
  import surface (`from drt.cli.main import _JsonFormatter` etc., used by tests)
  is unchanged.
- **Source-only connector docs** — add `docs/connectors/{duckdb,sqlite,redshift,
  sqlserver}.md`. These four sources were registered and listed in the README /
  drt-init skill but had no reference page (the drift check only maps
  destinations → docs). Now every registered connector has a docs page.

Verified: full unit suite 1832 passed / 5 skipped; ruff clean; **mypy across the
whole package** (141 files) clean — main.py's re-export resolves.
