+++ drt/cli/main.py
@@ -140,11 +140,13 @@ def _get_destination(sync: SyncConfig) -> Destination:
 # implementations now live in drt/cli/commands/run.py; re-export so the
 # legacy import path keeps working. New callers should import from
 # drt.cli.commands.run directly.
+from drt.cli._logging import (  # noqa: E402, F401
+    _configure_json_logging,
+    _JsonFormatter,
+)
 from drt.cli.commands.run import (  # noqa: E402, F401
     LogFormat,
-    _configure_json_logging,
     _exit_code_for_signal,
-    _JsonFormatter,
     _print_watermark_summary,
     _run_one,
     _RunContext,
